Ensure client does not block perpetually if the recipient does not send an ACK
Adds a timeout on the read operation of the stream when sending a message to a recipient. This ensures that we do not wait perpetually if the recipient never responds. Previously it would block on the stream.ReadByte() call. Now this will throw a SocketException after the timeout (default 30 second) elapses.